Add a new system test type, REUSEINITFILES
This test does two runs:

(1) A standard initial run

(2) A run that reuses the init-generated files from run (1).

This verifies that it works to reuse these init-generated files, and that you can get
bit-for-bit results by doing so. This is important because these files are typically
reused whenever a user reruns an initial case.

+++ CIME/SystemTests/system_tests_common.py
@@ -31,6 +31,9 @@ import glob, gzip, time, traceback, os
 
 logger = logging.getLogger(__name__)
 
+# Name of directory under the run directory in which init-generated files are placed
+INIT_GENERATED_FILES_DIRNAME = "init_generated_files"
+
 
 class SystemTestsCommon(object):
     def __init__(self, case, expected=None):
@@ -351,9 +354,20 @@ class SystemTestsCommon(object):
         self._case.load_env(reset=True)
         self._caseroot = case.get_value("CASEROOT")
 
-    def run_indv(self, suffix="base", st_archive=False, submit_resubmits=None):
+    def run_indv(
+        self,
+        suffix="base",
+        st_archive=False,
+        submit_resubmits=None,
+        keep_init_generated_files=False,
+    ):
         """
         Perform an individual run. Raises an EXCEPTION on fail.
+
+        keep_init_generated_files: If False (the default), we remove the
+        init_generated_files subdirectory of the run directory before running the case.
+        This is usually what we want for tests, but some specific tests may want to leave
+        this directory in place, so can set this variable to True to do so.
         """
         stop_n = self._case.get_value("STOP_N")
         stop_option = self._case.get_value("STOP_OPTION")
@@ -368,11 +382,16 @@ class SystemTestsCommon(object):
         for compout in glob.iglob(os.path.join(rundir, "*.cprnc.out")):
             os.remove(compout)
 
-        # remove all files in init_generated_files directory if it exists
-        init_generated_files_dir = os.path.join(rundir, "init_generated_files")
-        if os.path.isdir(init_generated_files_dir):
-            for init_file in glob.iglob(os.path.join(init_generated_files_dir, "*")):
-                os.remove(init_file)
+        if not keep_init_generated_files:
+            # remove all files in init_generated_files directory if it exists
+            init_generated_files_dir = os.path.join(
+                rundir, INIT_GENERATED_FILES_DIRNAME
+            )
+            if os.path.isdir(init_generated_files_dir):
+                for init_file in glob.iglob(
+                    os.path.join(init_generated_files_dir, "*")
+                ):
+                    os.remove(init_file)
 
         infostr = "doing an {:d} {} {} test".format(stop_n, stop_option, run_type)
 
@@ -839,7 +858,13 @@ class FakeTest(SystemTestsCommon):
                     )
                 )
 
-    def run_indv(self, suffix="base", st_archive=False, submit_resubmits=None):
+    def run_indv(
+        self,
+        suffix="base",
+        st_archive=False,
+        submit_resubmits=None,
+        keep_init_generated_files=False,
+    ):
         mpilib = self._case.get_value("MPILIB")
         # This flag is needed by mpt to run a script under mpiexec
         if mpilib == "mpt":
@@ -952,7 +977,13 @@ class TESTRUNFAILRESET(TESTRUNFAIL):
     rerun after an initial failure.
     """
 
-    def run_indv(self, suffix="base", st_archive=False, submit_resubmits=None):
+    def run_indv(
+        self,
+        suffix="base",
+        st_archive=False,
+        submit_resubmits=None,
+        keep_init_generated_files=False,
+    ):
         # Make sure STOP_N matches the original value for the case. This tests that STOP_N
         # has been reset properly if we are rerunning the test after a failure.
         env_test = EnvTest(self._get_caseroot())
